+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Control the server from server.js in the project.
 - Fix registration of user-defined Handlebars helpers.
 - Moved the Handlebars helper Hipsum.js to project dependencies.
+- The Handlebars instance is passed to project. Project template helpers are now registered at the project level instead of in the xtc module.
 
 ## 0.8.0-beta7
 - xtc-generator is now part of the xtc package and can be run from the project folder.